仪表板，视图删除确认提示；dashboard导出后刷新功能；新建仪表板无数据时提示

+++ src/page/dashboard/list/list.vue
@@ -162,7 +162,12 @@ export default {
     delDashboard(index,row){
       var _this = this;
       var url = '/api/show/dashboard/'+row.bid;
-      _this.$axios.delete(url,{}).then((res) => {
+      this.$confirm('确定删除此视图?', '提示', {
+        confirmButtonText: '确定',
+        cancelButtonText: '取消',
+        type: 'warning'
+      }).then(() => {
+        _this.$axios.delete(url,{}).then((res) => {
           if(res.data.code != 1){
             _this.$message({
               type: 'error',
@@ -177,13 +182,14 @@ export default {
             })
           }
           _this.requestData();
-      }).catch((err) => {
-          _this.$message({
-              type: 'error',
-              message: '网络错误，请重试',
-              showClose: true
-          })
-      })
+        }).catch((err) => {
+            _this.$message({
+                type: 'error',
+                message: '网络错误，请重试',
+                showClose: true
+            })
+        })
+      }).catch(() => {});
     },
     search(){
       this.page.currentPage = 1;
